fix->优化处理将持久化逻辑收敛到save动作中
